Refactor: Details: componentize in prep for pre-Details reuse
